+++ plugin-parent/common/src/main/java/jscover/maven/JasmineWebDriverRunner.java
@@ -10,6 +10,12 @@ import java.util.List;
 import static java.lang.String.format;
 
 public abstract class JasmineWebDriverRunner implements WebDriverRunner {
+    protected int timeOutSeconds;
+
+    public void setTimeOutSeconds(int timeOutSeconds) {
+        this.timeOutSeconds = timeOutSeconds;
+    }
+
     public List<String> getFailures(WebDriver webClient) {
         List<String> failures = new ArrayList<String>();
         List<WebElement> elements = webClient.findElements(By.className("failed"));