+++ plugin-parent/common/src/main/java/jscover/maven/WebDriverRunner.java
@@ -7,6 +7,7 @@ import org.openqa.selenium.WebDriver;
 import java.util.List;
 
 public interface WebDriverRunner {
+    void setTimeOutSeconds(int timeOutSeconds) throws MojoExecutionException;
     void waitForTestsToComplete(WebDriver webClient) throws MojoExecutionException;
     void verifyTestsPassed(WebDriver webClient) throws MojoFailureException;
     List<String> getFailures(WebDriver webClient);