+++ History.md
@@ -4,6 +4,7 @@
   * Add custom test runner support
   * Add clear message if `testDirectory` is not a sub-directory of `srcDir` (https://github.com/tntim96/JSCover/issues/146)
   * Clear localStorage before each run if using it (issue for PhantomJS)
+  * Make test timeout configurable
 
 1.0.5 / 2014-06-05
 ==================